refactor: rename sandbox utils module
(cherry picked from commit 7bd659cb92e1b198e1ef31fc5574c141faceaa11)
