+++ agentkit/toolkit/cli/sandbox/cli_exec.py
@@ -55,7 +55,7 @@ from agentkit.toolkit.cli.sandbox.tool_resolve import (
     find_tool_model_provider,
     get_remote_tool_model_provider,
 )
-from agentkit.toolkit.cli.sandbox.utils import (
+from agentkit.toolkit.cli.sandbox.sandbox_client import (
     add_session_terminal_shell_id,
     build_terminal_ws_url,
     error,
