+++ agentkit/toolkit/cli/sandbox/session_sync.py
@@ -24,7 +24,7 @@ from agentkit.toolkit.cli.sandbox.tool_resolve import (
     SandboxToolType,
     resolve_existing_sandbox_tool_id,
 )
-from agentkit.toolkit.cli.sandbox.utils import (
+from agentkit.toolkit.cli.sandbox.sandbox_client import (
     find_session_result,
     replace_tool_session_results,
 )
